rename folder and remove middleware
