+++ src/router/middleware/index.js
@@ -1,5 +1,4 @@
 import openInvoice from './openInvoice'
 import profileRequired from './profileRequired'
-import firstCreatingOrRecovering from './firstCreatingOrRecovering'
 
-export { openInvoice, profileRequired, firstCreatingOrRecovering }
+export { openInvoice, profileRequired }
